prebuilts: reorganize how we consume and publish
Prebuilts are now handled differently and require a proper release
process, in order to allow forks to also release we need to consume
GITHUB_REPOSITORY while rendering the repo files.

To build our binary and provide a prebuilt for those consuming over
bzlmod, then we need a better repository structure.

Change the release process, which is now heaviliy inspired in
https://github.com/aspect-build/bazel-lib which also provides a few
prebuilt golang binaries